test: adjust the running order
